phase-4/backend/pyproject.toml
[project]
name = "evolution-todo-backend"
version = "1.0.0"
description = "Evolution Todo API - Phase 4 Kubernetes Deployment"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Evolution Todo Team" }
]
keywords = ["fastapi", "todo", "kubernetes", "helm", "ai-chatbot"]

dependencies = [
    # Core API Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0.0",

    # Database & ORM
    "sqlmodel>=0.0.31",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.13.0",

    # Authentication & Security
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.21",

    # AI Chatbot (Phase 3) — openai-agents SDK (Agent + Runner pattern)
    "openai>=1.54.0",
    "openai-agents>=0.0.3",
    "openai-chatkit>=0.1.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",

    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",

    # Production Features
    "slowapi>=0.1.9",
    "cachetools>=5.3.0",
    "python-json-logger>=2.0.7",

    # Utilities
    "sse-starlette>=2.0.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=. --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
